+++ src/app/core/interceptors/http-interceptor.ts
@@ -8,6 +8,7 @@ import {
 import { inject } from "@angular/core";
 import { Observable, throwError } from "rxjs";
 import { catchError, finalize } from "rxjs/operators";
+import { HeroDialogService } from "../services/hero-dialog/hero-dialog-service";
 import { LoadingService } from "../services/loading/loading-service";
 
 export const HTTPInterceptor: HttpInterceptorFn = (
@@ -15,11 +16,15 @@ export const HTTPInterceptor: HttpInterceptorFn = (
   next: HttpHandlerFn,
 ): Observable<HttpEvent<any>> => {
   const loadingService = inject(LoadingService);
+  const dialogService = inject(HeroDialogService);
 
   loadingService.show();
 
   return next(request).pipe(
-    catchError((error: HttpErrorResponse) => throwError(() => error)),
+    catchError((error: HttpErrorResponse) => {
+      dialogService.openErrorDialog();
+      return throwError(() => error);
+    }),
     finalize(() => {
       loadingService.hide();
     }),